chatbot/requirements.txt
# LLM Integration (Groq API)
groq>=0.9.0

# Embeddings (Gemini - cached in MongoDB after first run)
google-genai>=1.0.0

# Embeddings / Vector Search
numpy==1.26.4

# Data Handling
python-dotenv==1.0.0

# API Backend
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
pymongo==4.6.0
